Fold recipient-guard simulate coverage into SimulateV1_MezoDivergence
The recipient guard is a mezod intentional deviation from geth's
eth_simulateV1 surface — it lives in the same conceptual bucket as the
existing per-call CanTransfer divergence the file already pins. A
separate suite added a fresh deployer fixture and forwarder lookup for
the same describe-level setup; the divergence file's outer before()
covers both with a single fixture call.

Move the two recipient-guard contexts (inner-CALL forwarder simulated,
top-level value transfer with validation omitted) into
SimulateV1_MezoDivergence, extend the outer fixture tags and helper
state to load the Forwarder, and drop the standalone file.
